Move demographics models to separate folder
