+++ src/main/java/org/sagebionetworks/bridge/models/demographics/DemographicUser.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.models.studies;
+package org.sagebionetworks.bridge.models.demographics;
 
 import java.util.Map;
