+++ src/main/java/org/sagebionetworks/bridge/models/demographics/DemographicValuesValidationConfiguration.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.models.studies;
+package org.sagebionetworks.bridge.models.demographics;
 
 import org.sagebionetworks.bridge.models.BridgeEntity;
 import org.sagebionetworks.bridge.validators.DemographicValuesValidator.DemographicValuesValidationType;
