+++ src/main/java/org/sagebionetworks/bridge/services/DemographicService.java
@@ -19,9 +19,9 @@ import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.accounts.Account;
 import org.sagebionetworks.bridge.models.appconfig.AppConfigElement;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicUser;
-import org.sagebionetworks.bridge.models.studies.DemographicValuesValidationConfiguration;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicUser;
+import org.sagebionetworks.bridge.models.demographics.DemographicValuesValidationConfiguration;
 import org.sagebionetworks.bridge.validators.DemographicUserValidator;
 import org.sagebionetworks.bridge.validators.DemographicValuesValidator;
 import org.sagebionetworks.bridge.validators.Validate;
