+++ src/main/java/org/sagebionetworks/bridge/validators/DemographicValuesValidator.java
@@ -10,9 +10,9 @@ import java.util.Set;
 
 import org.apache.commons.lang3.LocaleUtils;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
-import org.sagebionetworks.bridge.models.studies.DemographicValuesValidationConfiguration;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.DemographicValuesValidationConfiguration;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
