+++ src/test/java/org/sagebionetworks/bridge/dynamodb/DynamoParticipantVersionTest.java
@@ -24,8 +24,8 @@ import org.sagebionetworks.bridge.TestConstants;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 import org.sagebionetworks.bridge.models.accounts.ParticipantVersion;
 import org.sagebionetworks.bridge.models.accounts.SharingScope;
-import org.sagebionetworks.bridge.models.studies.Demographic;
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.Demographic;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
 
 public class DynamoParticipantVersionTest {
     private static final long CREATED_ON = DateTime.parse("2021-10-03T11:25:17.617-0700").getMillis();
