+++ src/test/java/org/sagebionetworks/bridge/validators/DemographicValueValidatorTest.java
@@ -8,7 +8,7 @@ import static org.testng.Assert.assertTrue;
 
 import java.math.BigDecimal;
 
-import org.sagebionetworks.bridge.models.studies.DemographicValue;
+import org.sagebionetworks.bridge.models.demographics.DemographicValue;
 import org.testng.annotations.Test;
 
 public class DemographicValueValidatorTest {
